requirements.txt
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
Deprecated==1.2.13
Django==3.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
joblib==1.2.0
jwcrypto==1.4.2
lightgbm==3.3.3
Markdown==3.4.1
mysqlclient==2.1.1
numpy==1.23.4
oauthlib==3.2.2
pycparser==2.21
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
sqlparse==0.4.3
threadpoolctl==3.1.0
urllib3==1.26.13
wrapt==1.14.1
zipp==3.10.0
celery==5.2.7
pandas==1.5.2
django-cleanup==6.0.0